# contract/openapi.json
{
  "components": {
    "schemas": {
...
      "ClipIn": {
        "additionalProperties": false,
        "description": "一次本地视频加工：按顺序裁出各段拼成一条，产物是本系统桶里的公开地址。\n\n``reference`` 是编辑时切给模型看的参考片段，只能在一条完整视频上裁一段，不重编码\n（起点因此落在最近的关键帧上，产物可能比区间略长）；``master`` 是拼出来的成片，各段\n参数互不相同，一律重编码对齐。两者存在不同前缀下，成片不进过期规则。",
        "properties": {
          "conversationId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversationid"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "purpose": {
            "enum": [
              "reference",
              "master"
            ],
            "title": "Purpose",
            "type": "string"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/ClipSegmentIn"
            },
            "maxItems": 50,
            "minItems": 1,
            "title": "Segments",
            "type": "array"
          },
          "taskId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Taskid"
          }
        },
        "required": [
          "purpose",
          "segments"
        ],
        "title": "ClipIn",
        "type": "object"
      },
      "ClipSegmentIn": {
        "additionalProperties": false,
        "description": "从一条视频里取 ``[start, end)`` 这一段，单位秒。",
        "properties": {
          "end": {
            "title": "End",
            "type": "number"
          },
          "start": {
            "minimum": 0.0,
            "title": "Start",
            "type": "number"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "start",
          "end"
        ],
        "title": "ClipSegmentIn",
        "type": "object"
      },
...
      "GenerationEnvelope": {
        "additionalProperties": false,
        "properties": {
          "generation": {
            "$ref": "#/components/schemas/GenerationOut"
          }
        },
        "required": [
          "generation"
        ],
        "title": "GenerationEnvelope",
        "type": "object"
      },
      "GenerationOut": {
        "additionalProperties": false,
        "description": "一次生成对外的样子。\n\n只给调用方用得上的：图在哪、跑到哪一步、失败了给人看什么。provider 名称、原始\n快照、租约与各段时间戳都是排队与排障的内部机制，快照里还带着 provider 的签名\nURL；来源对话不写回去——查的时候本来就是按它查的。",
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "errorMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errormessage"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "outputUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outputurl"
          },
          "request": {
            "additionalProperties": true,
            "title": "Request",
            "type": "object"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "taskId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Taskid"
          },
          "watermarkOutputUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Watermarkoutputurl"
          }
        },
        "required": [
          "id",
          "kind",
          "status",
          "request",
          "metadata",
          "taskId",
          "outputUrl",
          "watermarkOutputUrl",
          "errorMessage",
          "createdAt"
        ],
        "title": "GenerationOut",
        "type": "object"
      },
      "GenerationsPageOut": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/GenerationOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "GenerationsPageOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "description": "请求校验失败，与领域错误同一个信封。",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImageGenerationIn": {
        "additionalProperties": false,
        "description": "一次图像生成的输入。参考图为空即文生图，否则走图像编辑。",
        "properties": {
          "aspectRatio": {
            "enum": [
              "1:1",
              "3:2",
              "2:3",
              "3:4",
              "4:3",
              "4:5",
              "5:4",
              "9:16",
              "16:9",
              "21:9"
            ],
            "title": "Aspectratio",
            "type": "string"
          },
          "channel": {
            "anyOf": [
              {
                "enum": [
                  "dev",
                  "pro"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel"
          },
          "conversationId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversationid"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "model": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "prompt": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          },
          "referenceImageUrls": {
            "default": [],
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "title": "Referenceimageurls",
            "type": "array"
          },
          "resolution": {
            "default": "1k",
            "enum": [
              "1k",
              "2k",
              "4k"
            ],
            "title": "Resolution",
            "type": "string"
          },
          "taskId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Taskid"
          },
          "userName": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          }
        },
        "required": [
          "prompt",
          "aspectRatio"
        ],
        "title": "ImageGenerationIn",
        "type": "object"
      },
...
      "VideoGenerationIn": {
        "additionalProperties": false,
        "description": "一次视频生成的输入。字段照上游异步接口，外加归属字段、坐标 ``metadata`` 与结构化的 ``shot``。\n\n只拦本系统能判的：模型在允许表里（受理层）、地址是 http(s)、秒数不小于 -1、``shot``\n自身对得上（图片引用不越界、编号与正文一致）。画幅、分辨率、时长范围、素材规格由上游\n按模型判，这里不复制一份。",
        "properties": {
          "aspect_ratio": {
            "anyOf": [
              {
                "maxLength": 20,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aspect Ratio"
          },
          "conversation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "generate_audio": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generate Audio"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "model": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Model",
            "type": "string"
          },
          "prompt": {
            "anyOf": [
              {
                "maxLength": 4000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt"
          },
          "provider_options": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Options"
          },
          "reference_audio_urls": {
            "default": [],
            "items": {
              "type": "string"
            },
            "maxItems": 30,
            "title": "Reference Audio Urls",
            "type": "array"
          },
          "reference_image_urls": {
            "default": [],
            "items": {
              "type": "string"
            },
            "maxItems": 30,
            "title": "Reference Image Urls",
            "type": "array"
          },
          "reference_video_urls": {
            "default": [],
            "items": {
              "type": "string"
            },
            "maxItems": 30,
            "title": "Reference Video Urls",
            "type": "array"
          },
          "resolution": {
            "anyOf": [
              {
                "maxLength": 50,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution"
          },
          "seconds": {
            "anyOf": [
              {
                "minimum": -1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seconds"
          },
          "shot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VideoShotIn"
              },
              {
                "type": "null"
              }
            ]
          },
          "shot_index": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shot Index"
          },
          "task_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "user_name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Name"
          }
        },
        "required": [
          "model"
        ],
        "title": "VideoGenerationIn",
        "type": "object"
      },
...
      "VideoShotIn": {
        "additionalProperties": false,
        "description": "结构化的镜头组：全局设定加逐镜时间线。发给模型的正文由服务端按 shot_prompt 的规则拼。",
        "properties": {
          "global_settings": {
            "maxLength": 4000,
            "title": "Global Settings",
            "type": "string"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/VideoShotTimelineItemIn"
            },
            "minItems": 1,
            "title": "Timeline",
            "type": "array"
          }
        },
        "required": [
          "global_settings",
          "timeline"
        ],
        "title": "VideoShotIn",
        "type": "object"
      },
      "VideoShotTimelineItemIn": {
        "additionalProperties": false,
        "description": "镜头组里的一镜：起止时间、说什么、引了哪几张图。与分镜文件 video_shot.json 里的一镜同形。",
        "properties": {
          "image_indexes": {
            "items": {
              "minimum": 1.0,
              "type": "integer"
            },
            "title": "Image Indexes",
            "type": "array"
          },
          "prompt": {
            "maxLength": 4000,
            "title": "Prompt",
            "type": "string"
          },
          "timestamps": {
            "maxItems": 2,
            "minItems": 2,
            "prefixItems": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "minimum": 0.0,
                "type": "number"
              }
            ],
            "title": "Timestamps",
            "type": "array"
          }
        },
        "required": [
          "timestamps",
          "prompt",
          "image_indexes"
        ],
        "title": "VideoShotTimelineItemIn",
        "type": "object"
      },
      "VideoSubmitOut": {
        "additionalProperties": false,
        "description": "视频提交的回执，照上游：只有任务号。这里的 ``task_id`` 是我们这条生成记录的 id。",
        "properties": {
          "task_id": {
            "format": "uuid",
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "task_id"
        ],
        "title": "VideoSubmitOut",
        "type": "object"
      },
      "VideoTaskError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "VideoTaskError",
        "type": "object"
      },
      "VideoTaskOut": {
        "additionalProperties": false,
        "description": "视频任务快照，照上游任务查询的形状。",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VideoTaskError"
              },
              {
                "type": "null"
              }
            ]
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VideoTaskResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "format": "uuid",
            "title": "Task Id",
            "type": "string"
          },
          "type": {
            "const": "video",
            "default": "video",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "status",
          "created_at"
        ],
        "title": "VideoTaskOut",
        "type": "object"
      },
      "VideoTaskResult": {
        "additionalProperties": false,
        "properties": {
          "output_url": {
            "title": "Output Url",
            "type": "string"
          },
          "watermark_output_url": {
            "title": "Watermark Output Url",
            "type": "string"
          }
        },
        "required": [
          "output_url",
          "watermark_output_url"
        ],
        "title": "VideoTaskResult",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "in": "cookie",
        "name": "iclip_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "iclip-agent",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/generations": {
      "get": {
        "description": "给了 ``conversationId`` / ``taskId`` 就只列那段对话、那张需求单下面的记录，可见性口径不变。\n\n``metadata`` 在查询串里是一段 JSON 对象，按包含匹配筛（分镜页拿它按镜头组、按帧查）。",
        "operationId": "list_generations_generations_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "conversationId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Conversationid"
            }
          },
          {
            "in": "query",
            "name": "taskId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Taskid"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "image",
                    "video",
                    "clip"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "description": "JSON 对象；只列坐标包含这些键值的记录，服务端不解释键的含义",
            "in": "query",
            "name": "metadata",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "JSON 对象；只列坐标包含这些键值的记录，服务端不解释键的含义",
              "title": "Metadata"
            }
          },
          {
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationsPageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Generations"
      }
    },
    "/generations/clips": {
      "post": {
        "description": "提交一次本地视频加工：按 ``segments`` 的顺序裁出各段拼成一条，产物存进本系统的桶。\n\n``purpose=reference`` 是编辑时切给模型看的参考片段，只能在一条完整视频上裁一段、\n不重编码；``purpose=master`` 是拼出来的成片，一律重编码对齐参数。不经外部服务，\n不计费，也没有 ``userName``。",
        "operationId": "submit_clip_generations_clips_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClipIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Clip"
      }
    },
    "/generations/image": {
      "post": {
        "description": "提交一次图片生成。``userName`` 的规则与视频相同。",
        "operationId": "submit_image_generations_image_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageGenerationIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Image"
      }
    },
...
    "/generations/video": {
      "post": {
        "description": "提交一次视频生成。请求体照上游异步接口，外加 conversation_id / task_id / metadata。\n\n正文可以直接给 ``prompt``，也可以给结构化的 ``shot`` 由服务端拼成 ``prompt``；记录里\n两者都存，``shot`` 不发上游。``user_name``：API key 调用方必填、照收；浏览器会话可\n省略，填登录用户名。",
        "operationId": "submit_video_generations_video_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoGenerationIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoSubmitOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Video"
      }
    },
...
    "/generations/video/{task_id}": {
      "get": {
        "description": "视频任务快照，照上游任务查询的形状。只认视频记录，可见性与 ``GET /generations/{id}`` 相同。",
        "operationId": "get_video_task_generations_video__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoTaskOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Video Task"
      }
    },
    "/generations/{job_id}": {
      "get": {
        "operationId": "get_generation_generations__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Generation"
      }
    },
...
  }
}